requirements.txt
numpy==1.21.4
flask==2.0.2
tensorflow==2.7.0
pillow==8.4.0